Fix #94, mote stuff being ints that should be longs

+++ Common/src/main/java/ram/talia/hexal/api/spell/iota/MoteIota.java
@@ -108,7 +108,7 @@ public class MoteIota extends Iota {
         return MediafiedItemManager.typeMatches(this.getItemIndex(), other);
     }
 
-    public @Nullable MoteIota splitOff(int amount, @Nullable UUID storage) {
+    public @Nullable MoteIota splitOff(long amount, @Nullable UUID storage) {
         var newIndex = MediafiedItemManager.splitOff(this.getItemIndex(), amount, storage);
         if (newIndex == null)
             return null;
